Refactoring refresh and access tokens flow 1.0 #4
