+++ .prettierrc
@@ -1,3 +1,4 @@
 {
-  "singleQuote": true
+    "singleQuote": true,
+    "tabWidth": 4
 }
